(PUP-7864) Move tests related to Task loader move tests to task_spec
This commit reverts the loaders/module_loaders_spec.rb to the state that
it has on the master branch and moves all tests related to `Task`,
`GenericTask` and the loading of tasks to the types/task_spec.rb.
